requirements.txt
pandas==1.3.2
Pillow==9.5.0
streamlit==1.21.0
